added screenshots folder
